http://download.virtualbox.org/virtualbox/4.3.0_BETA1/VirtualBox-4.3.0_BETA1.tar.bz2

Merge commit '59db47683cc02beebcf10cdb1625d23bbe6ba294' of git://github.com/larsr/VirtualBox-OSE into VirtualBox/4.3/tar

Conflicts:
	Config.kmk
	include/VBox/com/MultiResult.h
	include/VBox/vmm/pdmcardreaderinfs.h
	include/VBox/vmm/pdmnetshaperint.h
	src/VBox/Additions/WINNT/Graphics/Video/disp/wddm/VBoxDispD3D64.def
	src/VBox/Additions/WINNT/Graphics/Video/disp/wddm/VBoxDispMpTst.cpp
	src/VBox/Additions/WINNT/Graphics/Video/disp/wddm/VBoxScreen.cpp
	src/VBox/Additions/WINNT/Graphics/Wine/switcher/sw_common.c
	src/VBox/Additions/common/VBoxService/VBoxServiceControlThread.cpp
	src/VBox/Devices/Audio/DevCodec.cpp
	src/VBox/Devices/Audio/DevCodec.h
	src/VBox/Devices/Audio/DevIchIntelHDA.cpp
	src/VBox/Devices/EFI/DevOVMF.cpp
	src/VBox/Devices/EFI/Firmware2/VBoxPkg/VBoxFsDxe/hfs_format.h
	src/VBox/Devices/EFI/Firmware2/VBoxPkg/VBoxPkg.dsc
	src/VBox/Devices/EFI/Firmware2/VBoxPkg/VBoxPkg.fdf
	src/VBox/Devices/EFI/Firmware2/VBoxPkg/VBoxPkgX64.dsc
	src/VBox/Devices/Graphics/BIOS-new/Makefile.kmk
	src/VBox/Devices/Graphics/BIOS-new/VBoxVgaBiosAlternative.asm
	src/VBox/Devices/Network/slirp/dnsproxy/dnsproxy.c
	src/VBox/Devices/PC/DevPcArch.c
	src/VBox/Devices/PC/Etherboot-src/Makefile.kmk
	src/VBox/Devices/PC/Etherboot-src/arch/i386/drivers/net/undi.c
	src/VBox/Devices/PC/Etherboot-src/arch/i386/drivers/net/undi.h
	src/VBox/Devices/PC/Etherboot-src/arch/i386/firmware/pcbios/basemem.c
	src/VBox/Devices/PC/Etherboot-src/arch/i386/include/callbacks_arch.h
	src/VBox/Devices/PC/Etherboot-src/core/dns_resolver.c
	src/VBox/Devices/PC/Etherboot-src/include/callbacks.h
	src/VBox/Devices/PC/Etherboot-src/include/pxe.h
	src/VBox/Devices/Storage/ATAController.cpp
	src/VBox/Devices/Storage/ATAController.h
	src/VBox/Devices/Storage/fdc.c
	src/VBox/Frontends/VBoxAutostart/VBoxAutostart.cpp
	src/VBox/Frontends/VBoxBFE/DisplayImpl.cpp
	src/VBox/Frontends/VBoxBFE/DisplayImpl.h
	src/VBox/Frontends/VBoxBFE/KeyboardImpl.cpp
	src/VBox/Frontends/VBoxBFE/KeyboardImpl.h
	src/VBox/Frontends/VBoxBFE/Makefile.kmk
	src/VBox/Frontends/VBoxBFE/StatusImpl.cpp
	src/VBox/Frontends/VBoxBFE/StatusImpl.h
	src/VBox/Frontends/VBoxBFE/VBoxBFE.cpp
	src/VBox/Frontends/VBoxBFE/VBoxBFEHardened.cpp
	src/VBox/Frontends/VBoxBFE/VMMDev.h
	src/VBox/Frontends/VBoxBFE/VMMDevInterface.cpp
	src/VBox/Frontends/VBoxBFE/testcase/tstMouseImpl.cpp
	src/VBox/Frontends/VBoxHeadless/VideoCapture/Makefile.kmk
	src/VBox/Frontends/VirtualBox/src/VBoxFBQGL.cpp
	src/VBox/Frontends/VirtualBox/src/runtime/UIFrameBufferDirectDraw.cpp
	src/VBox/Frontends/VirtualBox/src/runtime/UIFrameBufferQGL.cpp
	src/VBox/Frontends/VirtualBox/src/runtime/UIFrameBufferSDL.cpp
	src/VBox/Frontends/VirtualBox/src/runtime/UIMachineShortcuts.cpp
	src/VBox/Frontends/VirtualBox/src/selector/UISelectorShortcuts.cpp
	src/VBox/Frontends/VirtualBox/src/selector/UIVMListView.cpp
	src/VBox/Frontends/VirtualBox/src/selector/UIVMPreviewWindow.cpp
	src/VBox/Frontends/VirtualBox/src/widgets/VBoxMiniToolBar.cpp
	src/VBox/Frontends/VirtualBox/src/widgets/VBoxWarningPane.cpp
	src/VBox/HostDrivers/Support/SUPDrvInternal.h
	src/VBox/Installer/solaris/vboxconfig.sh
	src/VBox/Main/include/DHCPServerRunner.h
	src/VBox/Main/include/ProgressCombinedImpl.h
	src/VBox/Main/src-client/GuestSessionImplTasks.cpp
	src/VBox/Main/src-client/UsbCardReader.cpp
	src/VBox/Main/src-server/DHCPServerRunner.cpp
	src/VBox/NetworkServices/NAT/VBoxNetNAT.cpp
	src/VBox/Runtime/r3/os2/poll-os2.cpp
	src/VBox/Runtime/r3/posix/poll-posix.cpp
	src/VBox/Runtime/r3/win/poll-win.cpp
	src/VBox/Runtime/testcase/ioctl.h
	src/VBox/Runtime/testcase/tstIoCtl.cpp
	src/VBox/Runtime/testcase/tstMp-1.cpp
	src/VBox/Runtime/testcase/tstTime.cpp
	src/VBox/Storage/testcase/tstVDCompact.vd
	src/VBox/Storage/testcase/tstVDCopy.vd
	src/VBox/Storage/testcase/tstVDDiscard.vd
	src/VBox/Storage/testcase/tstVDIo.vd
	src/VBox/Storage/testcase/tstVDShareable.vd
	src/VBox/VMM/VMMAll/HWACCMAll.cpp
	src/VBox/VMM/VMMR0/HWACCMR0.cpp
	src/VBox/VMM/VMMR0/HWACCMR0A.asm
	src/VBox/VMM/VMMR0/HWACCMR0Mixed.mac
	src/VBox/VMM/VMMR0/HWSVMR0.cpp
	src/VBox/VMM/VMMR0/HWSVMR0.h
	src/VBox/VMM/VMMR0/HWVMXR0.cpp
	src/VBox/VMM/VMMR0/HWVMXR0.h
	src/VBox/VMM/VMMR0/PDMNetShaperR0.cpp
	src/VBox/VMM/VMMR3/DBGFSym.cpp
	src/VBox/VMM/VMMR3/EMHwaccm.cpp
	src/VBox/VMM/VMMR3/HWACCM.cpp
	src/VBox/VMM/VMMRC/HWACCMRCA.asm
	src/VBox/VMM/include/HWACCMInternal.h
	src/VBox/VMM/include/HWACCMInternal.mac
	src/VBox/VMM/include/internal/em.h
	src/VBox/VMM/include/internal/vm.h
	src/VBox/VMM/testcase/tstVMM-HwAccm.cpp

+++ src/VBox/Main/glue/string.cpp
@@ -1,10 +1,10 @@
-/* $Id: string.cpp 42570 2012-08-03 09:56:09Z vboxsync $ */
+/* $Id: string.cpp $ */
 /** @file
  * MS COM / XPCOM Abstraction Layer - UTF-8 and UTF-16 string classes.
  */
 
 /*
- * Copyright (C) 2006-2011 Oracle Corporation
+ * Copyright (C) 2006-2012 Oracle Corporation
  *
  * This file is part of VirtualBox Open Source Edition (OSE), as
  * available from http://www.virtualbox.org. This file is free software;
@@ -57,7 +57,7 @@ void Bstr::copyFromN(const char *a_pszSrc, size_t a_cchMax)
         throw std::bad_alloc();
     }
 
-    m_bstr = ::SysAllocStringByteLen(NULL, cwc * sizeof(OLECHAR));
+    m_bstr = ::SysAllocStringByteLen(NULL, (unsigned)(cwc * sizeof(OLECHAR)));
     if (RT_UNLIKELY(!m_bstr))
         throw std::bad_alloc();
